+++ templates/header.tera.sh
@@ -1,0 +1,49 @@
+#!/usr/bin/env bash
+
+BASH_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
+
+cd "${BASH_DIR}"
+
+#----------------------------#
+# Colors in term
+#----------------------------#
+GREEN=
+RED=
+NC=
+if tty -s < /dev/fd/1 2> /dev/null; then
+    GREEN='\033[0;32m'
+    RED='\033[0;31m'
+    NC='\033[0m' # No Color
+fi
+
+log_warn () {
+    echo >&2 -e "${RED}==> $@ <==${NC}"
+}
+
+log_info () {
+    echo >&2 -e "${GREEN}==> $@${NC}"
+}
+
+log_debug () {
+    echo >&2 -e "==> $@"
+}
+
+#----------------------------#
+# helper functions
+#----------------------------#
+set +e
+
+# set stacksize to unlimited
+if [[ "$OSTYPE" != "darwin"* ]]; then
+    ulimit -s unlimited
+fi
+
+signaled () {
+    log_warn Interrupted
+    exit 1
+}
+trap signaled TERM QUIT INT
+
+readlinkf () {
+    perl -MCwd -l -e 'print Cwd::abs_path shift' "$1";
+}